+++ src/pious/_executables/flops.py
@@ -1,32 +1,32 @@
-from argparse import Namespace, _SubParsersAction
+import typer
+from typing_extensions import Annotated
+from pathlib import Path
 from ..flops import Flops
 
 
-def exec_flops(args: Namespace):
+def flops_cmd(
+    filter_expr: Annotated[
+        str,
+        typer.Argument(
+            help="Filter flops based on texture, e.g., 'not flush and straight'"
+        ),
+    ] = None,
+    count: Annotated[
+        bool, typer.Option(help="Print number of flops satisfying filter")
+    ] = False,
+    file: Annotated[Path, typer.Option(help="Output file path")] = None,
+):
+    """Filter and print flop subsets"""
     flops = Flops()
 
-    if args.filter is not None:
-        flops.filter(args.filter)
-    print(flops)
-    if args.count:
-        print(f'{len(flops)} flops satisfy filter "{args.filter}"')
-    if args.file is not None:
-        with open(args.file, "w+") as f:
+    if filter_expr is not None:
+        flops.filter(filter_expr)
+    typer.echo(str(flops))
+    if count:
+        if filter_expr is None:
+            typer.echo(f"{len(flops)} flops in total")
+        else:
+            typer.echo(f'{len(flops)} flops satisfy filter "{filter_expr}"')
+    if file is not None:
+        with open(file, "w+") as f:
             f.write(str(flops))
-
-
-def register_command(sub_parsers: _SubParsersAction):
-    parser_flops = sub_parsers.add_parser(
-        "flops", description="Utility module for filtering and printing flop subsets"
-    )
-    parser_flops.set_defaults(function=exec_flops)
-
-    parser_flops.add_argument(
-        "filter",
-        nargs="?",
-        help="Filter flops based on texture, e.g., 'not flush and straight'",
-    )
-    parser_flops.add_argument(
-        "--count", action="store_true", help="Print number of flops satisfying filter"
-    )
-    parser_flops.add_argument("--file", "-f", help="print flops to file")